+++ src/main/java/io/github/tr100000/codec2schema/api/Codec2SchemaPlugin.java
@@ -1,5 +1,6 @@
 package io.github.tr100000.codec2schema.api;
 
+import com.google.gson.JsonObject;
 import net.minecraft.resources.Identifier;
 import org.jspecify.annotations.Nullable;
 
@@ -11,6 +12,8 @@ public interface Codec2SchemaPlugin {
     default boolean shouldRun() {
         return true;
     }
+
+    default void writeExtraInfo(JsonObject pluginInfoJson) {}
 
     default @Nullable Identifier getId() {
         return null;