+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-**Update to 26.3-snapshot-6**
+### Added
+- Added an option to export plugin info (ids, versions, exported schemas, etc.)
+
+### Changed
+- The game version in exported info now uses the internal name instead of the display name